refactor(ewm-service): rewrite and extend Statn-client module
