+++ stats/stat-client/src/main/java/ru/practicum/yandex/WebConfig.java
@@ -10,11 +10,8 @@ import org.springframework.web.reactive.function.client.WebClient;
 @Configuration
 public class WebConfig {
 
-    @Value("${stat-server.url}")
-    private String addressBaseUrl;
-
     @Bean
-    public WebClient webClient() {
+    public WebClient webClient(@Value("${stat-server.url}") String addressBaseUrl) {
         return WebClient.builder()
                 .baseUrl(addressBaseUrl)
                 .defaultHeader(HttpHeaders.CONTENT_TYPE, MediaType.APPLICATION_JSON_VALUE,
