+++ stats/stat-dto/src/main/java/ru/practicum/yandex/dto/EndpointHitDto.java
@@ -2,15 +2,16 @@ package ru.practicum.yandex.dto;
 
 import com.fasterxml.jackson.annotation.JsonFormat;
 import lombok.AllArgsConstructor;
+import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
-import ru.practicum.yandex.validation.ValidIPv4;
 
 import java.time.LocalDateTime;
 
 @Data
 @NoArgsConstructor
 @AllArgsConstructor(staticName = "of")
+@Builder
 public class EndpointHitDto {
 
     private Long id;
@@ -19,7 +20,6 @@ public class EndpointHitDto {
 
     private String uri;
 
-    @ValidIPv4
     private String ip;
 
     @JsonFormat(pattern = "yyyy-MM-dd HH:mm:ss")
